This does de-duplication of the `ProcessingAttempts` based on `TimeOfFailure` value. This is stored to the microsecond so a chance that we will have two failures for the same message happening at the same time (to the microsecond) is slim.

That said there are are two (a least) additional angles to this change:
  * de-duplication based on `TimeOfFailure` is a hack. I think we should introduce a proper `ProcessingAttemptId`
  * I'm not an expert in Raven but my hunch is that Javascript patch might be significantly slower than non-Javascript patch. I propose to do some testing around that

Thoughts?