@mikeminutillo 2020-09-08
> I think we should introduce a proper `ProcessingAttemptId`
:+1: :100:

Agreed. I would go further and make that the primary key and store each processing attempt separately. That way we have a forward-only state machine. `Failed` => `Archived` || `Retried`. Anything else we do to show prior attempts or to aggregate that data in any way becomes a display concern.