@SzymonPobiega 2020-09-08
> That way we have a forward-only state machine

❤️ Loops in state machines are the root of all evil