@tmasternak 2020-09-09
@mikeminutillo after doing some further analysis with Szymon and Sean we decided that moving `ProcessingAttempt` out of the `FailedMessage` is far from trivial:
* There are quite a few indexes that we would need to change and test,
* There are some fields in the `FailedMessage` that require consistency with attempt e.g. `Status` and `Count`. This would mean that we would need to do 2 writes first for the `ProcessingAttempt` and if the document is not duplicate we could update the `FailedMessage` properly,
* We considered the possibility of having `LatestProcessingAttempt` on the `FailedMesssage` but that left us still with 2 writes
* Finally we tried to figure out if we can store the `Attempts` in the `FailedMessage` as a map and have a patch that would update the map based on the insert key (`AttemptedAt`) but it looks that RavenDB does not support that.

In the end, we decided to go with the scripted patch approach. I'm doing performance tests and will update this issue with data.