@tmasternak 2020-09-09
### Rabbit Performance Test
Local RabbitMQ instance, 10K failed attempts in each test 

|Attempts per msg| ScriptedPatch [msg/sec)]| Current [msg/sec]|
|---|--|-------|
|2 | 38| 37|
|5| 35| 44|
|10K (all attempts for single doc) | 50 msg/sec | N/A|