@SzymonPobiega 2020-09-09
I don't see any performance issues here 🙈