@tmasternak 2020-09-09
Performance tests on ASBS transport do not show any performance difference.