@mikeminutillo 2020-09-10
I'm happy with those perf results. Most ingested error messages should have a single processing attempt and we don't see an appreciable difference when that holds true. In environments where there a significantly more processing attempts (and the table shows up to 5), going a little slower but potentially storing less data is acceptable. The times when we get large batches of errors with new processing attempts are likely to be after a ServiceControl initiated bulk-retry, at which time the server may be busy managing that process anyway.