@ramonsmits 2020-09-10
> > I think we should introduce a proper `ProcessingAttemptId`
> 
> 👍 💯
> 
> Agreed. I would go further and make that the primary key and store each processing attempt separately. That way we have a forward-only state machine. `Failed` => `Archived` || `Retried`. Anything else we do to show prior attempts or to aggregate that data in any way becomes a display concern.

Exactly what I said to @tmasternak, why do we expand a document? The current design isn't good for queurying, paging, etc. and we should have a CQRS type of design in a future version to deal with those indexes.

Not even having deduplication as a key but having duplicate makes it possible to do any kind of deduplication on the query side if needed or alternatively generate a unique key and start storing the attempts separately with that ID and add the ID in the attempts collection.

I didn't' expect any throughput performance issues simply because the operation is IO bound and any javascript CPU increase will not increase IO but potentially only CPU so a better performance test would be comparing CPU usage although I'm not really expecting a significant increase.