@ramonsmits 2020-09-10
By the way, microseconds is a hack but physically (yet...) impossible to do multiple retries that start in the same microsecond.